+++ GnuClay.Engine/ScriptExecutor/CommonData/IRunTimeTypeProcessingContext.cs
@@ -8,7 +8,7 @@ namespace GnuClay.Engine.ScriptExecutor.CommonData
 {
     public interface IRunTimeTypeProcessingContext
     {
-        IValue CreateValue(int typeKey, object value);
-        IValue CallBinaryOperator(int operatorKey, IValue firstOperand, IValue secondOperand);
+        IValue CreateValue(ulong typeKey, object value);
+        IValue CallBinaryOperator(ulong operatorKey, IValue firstOperand, IValue secondOperand);
     }
 }
